backend/requirements.txt
google-cloud-aiplatform[agent_engines]>=1.60.0
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0